Crash logs a warning with exception when starting in disabled mode

+++ SDK/AppCenterCrashes/Microsoft.AppCenter.Crashes.Windows.Shared/Utils/ErrorLogHelper.cs
@@ -307,17 +307,23 @@ namespace Microsoft.AppCenter.Crashes.Utils
         {
             lock (LockObject)
             {
-                AppCenterLog.Debug(Crashes.LogTag, $"Deleting error log directory.");
-                try
+                _crashesDirectory.Refresh();
+
+                // Crashes directory will not exist if disabling has already deleted the folder.
+                if (_crashesDirectory.Exists())
                 {
-                    _crashesDirectory.Delete(true);
-                }
-                catch (System.Exception ex)
-                {
-                    AppCenterLog.Warn(Crashes.LogTag, $"Failed to delete error log directory.", ex);
+                    AppCenterLog.Debug(Crashes.LogTag, $"Deleting error log directory.");
+                    try
+                    {
+                        _crashesDirectory.Delete(true);
+                    }
+                    catch (System.Exception ex)
+                    {
+                        AppCenterLog.Warn(Crashes.LogTag, $"Failed to delete error log directory.", ex);
+                    }
+                    AppCenterLog.Debug(Crashes.LogTag, "Deleted crashes local files.");
                 }
             }
-            AppCenterLog.Debug(Crashes.LogTag, "Deleted crashes local files.");
         }
 
         private static ModelException CreateModelException(System.Exception exception)
